Renamed project to com.infinyon.fluvio
