+++ build.rs
@@ -21,10 +21,10 @@ fn main() {
         .join("src")
         .join("main")
         .join("java")
-        .join("fluvio")
-        .join("client")
-        .join("java"),
-        "fluvio.client.java".into(), // TODO: Change to "com.infinyon.fluvio"
+        .join("com")
+        .join("infinyon")
+        .join("fluvio"),
+        "com.infinyon.fluvio".into(),
     )
     //.use_optional_package("com.hadisatrio.optional".into())
     .use_reachability_fence(if have_java_9 {
