+++ lib/src/main/java/com/infinyon/fluvio/Library.java
@@ -1,7 +1,7 @@
 /*
  * This Java source file was generated by the Gradle 'init' task.
  */
-package fluvio.client.java;
+package com.infinyon.fluvio;
 
 public class Library {
     static {
