+++ lib/src/test/java/com/infinyon/fluvio/LibraryTest.java
@@ -1,7 +1,7 @@
 /*
  * This Java source file was generated by the Gradle 'init' task.
  */
-package fluvio.client.java;
+package com.infinyon.fluvio;
 
 import org.junit.Test;
 import static org.junit.Assert.*;
